refactor: [FE] 타이머 컴포넌트의 시간측정 관련 로직을 훅으로 분리
- Intent
결과페이지에서 재사용하기 위해 분리
